Remove `allow`s and clean things up
