+++ frontend/exporter/src/lib.rs
@@ -6,10 +6,6 @@
 #![feature(macro_metavar_expr)]
 #![feature(if_let_guard)]
 #![feature(let_chains)]
-#![allow(unused_imports)]
-#![allow(unused_variables)]
-#![allow(unreachable_code)]
-#![allow(dead_code)]
 #![allow(incomplete_features)]
 #![feature(specialization)]
